# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Tenant RAG & Jev AI Service
  description: Multi-tenant knowledge indexing, hybrid retrieval and grounded generation. Tenant comes from the trusted context;
    Jev orchestrates capability only.
  version: 1.0.0
paths:
...
  /api/v1/settings:
    get:
      tags:
      - settings
      summary: Read Settings
      description: Masked current configuration (env + stored overrides).
      operationId: read_settings_api_v1_settings_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Tenant-Context
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Context
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Read Settings Api V1 Settings Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - settings
      summary: Update Settings
      description: Persist overrides, apply them to the live process, and report what changed.
      operationId: update_settings_api_v1_settings_put
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Tenant-Context
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Update Settings Api V1 Settings Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/settings/llm/models:
    post:
      tags:
      - settings
      summary: List Llm Models
      description: Ask an OpenAI-compatible endpoint which models it serves.
      operationId: list_llm_models_api_v1_settings_llm_models_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Tenant-Context
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelsProbeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Llm Models Api V1 Settings Llm Models Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/settings/jev/probe:
    post:
      tags:
      - settings
      summary: Probe Jev
      description: 'One cheap decision question: does this Jev endpoint answer, and how fast?'
      operationId: probe_jev_api_v1_settings_jev_probe_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Tenant-Context
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JevProbeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Probe Jev Api V1 Settings Jev Probe Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/settings/api-keys:
    get:
      tags:
      - settings
      summary: List Api Keys
      description: Daftar kunci yang bisa memanggil layanan ini. Nilai kunci tidak pernah ikut.
      operationId: list_api_keys_api_v1_settings_api_keys_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Tenant-Context
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Context
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Api Keys Api V1 Settings Api Keys Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - settings
      summary: Create Api Key
      description: Buat kunci baru. Nilai kunci dikembalikan **sekali** di sini dan tidak disimpan apa adanya.
      operationId: create_api_key_api_v1_settings_api_keys_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Tenant-Context
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Api Key Api V1 Settings Api Keys Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/settings/api-keys/{key_id}:
    delete:
      tags:
      - settings
      summary: Revoke Api Key
      description: Cabut kunci dari registry. Kunci dari env ditolak di sini (dikelola lewat API_KEYS_JSON).
      operationId: revoke_api_key_api_v1_settings_api_keys__key_id__delete
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
          title: Key Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Tenant-Context
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Context
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Revoke Api Key Api V1 Settings Api Keys  Key Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/settings/access/sessions:
    delete:
      tags:
      - settings
      summary: Revoke All Sessions
      description: Keluarkan semua sesi konsol, termasuk yang sekarang (kecuali diminta menyisakan satu).
      operationId: revoke_all_sessions_api_v1_settings_access_sessions_delete
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Tenant-Context
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Context
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Revoke All Sessions Api V1 Settings Access Sessions Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    JevProbeRequest:
      properties:
        provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        api_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Api Key
      additionalProperties: false
      type: object
      title: JevProbeRequest
...
    ModelsProbeRequest:
      properties:
        base_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Base Url
        api_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Api Key
      additionalProperties: false
      type: object
      title: ModelsProbeRequest
...
    SettingsUpdateRequest:
      properties:
        llm:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Llm
        jev:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Jev
        uploads:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Uploads
      additionalProperties: false
      type: object
      title: SettingsUpdateRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError